# Simplify and optimize Postgres query for primary_keys()