@maclover7 2017-02-09
@jordanlewis Can you take a look at the failing sqlite3 tests when you get a chance? Thanks!