@jordanlewis 2017-02-09
@maclover7 thanks for the pointer. The new test that I just added is failing - it looks like sqlite3 doesn't preserve the order of composite primary keys, like in @kamipo's example.

This raises the question: does ActiveRecord really expect and require the order of the output of `primary_keys()` to match the order of the composite key? It seems like perhaps it doesn't matter, if sqlite3 hasn't ever properly implemented this behavior.

If that's the case, then I'll drop the test and remove the `unnest()` from my patch, as it will no longer be necessary. Please advise!