@kamipo 2017-02-09
Originally I contained the reverse order test in #21614 and passed the test for all adapters in my local. But for some reason only failed sqlite3 adapter on travis. Debug on travis is hard to me, so I decided to drop the test in that time.

Key column order (include primary key) is very important, it should be kept.

I suggest to skip the reverse order test for sqlite3 adapter until the cause of failure on travis is found.