@jordanlewis 2017-02-09
Yeah, I am noticing the same thing - the tests are passing against sqlite locally. I'll try to track this down a little further, but I may end up skipping the test as you suggest.