@jordanlewis 2017-02-09
Ready for another look - I've marked the test as skipped under the sqlite3 adapter.